+++ NEWS.md
@@ -1,4 +1,5 @@
-# modelStudio (development)
+# modelStudio 2.1.0
+* **DEFAULTS CHANGES**: if `new_observation = NULL` then choose `new_observation_n = 3` observations, evenly spread by the histogram bins of `y_hat`. This shall always include the observations, which ids are `which.min(y_hat)` and `which.max(y_hat)`. Additionally, improve the observation dropdown text in dashboard. [(#94)](https://github.com/ModelOriented/modelStudio/issues/94)
 * This version requires `DALEX v2.0.1`
 * added new options to `ms_options`: `ms_subtitle`, `ms_margin_top` and `ms_margin_bottom`
 * added new parameters to `modelStudio()`: `N_fi = 10*N` and `B_fi = B`